Refactor git-metadata/utils
