+++ src/helpers/logger.ts
@@ -1,12 +1,5 @@
 import chalk from 'chalk'
 
-export const timedExecAsync = async <I, O>(f: (input: I) => Promise<O>, input: I): Promise<number> => {
-  const initialTime = Date.now()
-  await f(input)
-
-  return (Date.now() - initialTime) / 1000
-}
-
 export enum LogLevel {
   DEBUG = 1,
   INFO,
