+++ src/helpers/utils.ts
@@ -319,3 +319,10 @@ export const filterAndFormatGithubRemote = (rawRemote: string | undefined): stri
 
   return rawRemote
 }
+
+export const timedExecAsync = async <I, O>(f: (input: I) => Promise<O>, input: I): Promise<number> => {
+  const initialTime = Date.now()
+  await f(input)
+
+  return (Date.now() - initialTime) / 1000
+}
